Add files via upload
The index.html file is the main HTML file for a Weather Web App. It usually contains the structure and content for the web page, including the header, footer, and body of the page.The file contains elements such as a search bar for the user to input a location, a div for displaying the current weather conditions, and perhaps a forecast for the next few days, also includes links to other files such as CSS stylesheets and JavaScript scripts that are used to enhance the user experience or provide additional functionality to the app.