Add files via upload
The script.js file is a JavaScript file that is used to provide additional functionality to a Weather Web App. It is linked to from the index.html file and is executed by the web browser when the page is loaded.  script.js file contains functions for making API calls to a weather service to retrieve data about the current weather conditions and forecast for a given location. It is responsible for handling user input, icluding  when the user submits a location for which they want to see the weather. Additionally, the script.js file contains code for updating the DOM (Document Object Model) with the data retrieved from the API, in order to display the weather information to the user.