Add files via upload
A style.css file is a stylesheet that specifies the layout and formatting of a web page written in HTML or other markup language. It is a separate file that is linked to a web page and contains all of the styles and formatting rules for the layout and of the app's user interface, including elements such as the font, color, and size of text, the placement and appearance of buttons and other interactive elements, and the overall design of the app. By separating the style rules from the content of the web page, the style.css file makes it easier to maintain and update the design of the app without having to make changes to the underlying HTML code.